move units from info/ to shared/
