+++ frontend/src/modules/alarms/limits/AlarmLimitsPage.tsx
@@ -15,7 +15,7 @@ import { VentilationMode } from '../../../store/proto/mcu_pb';
 import { getParametersRequestMode } from '../../../store/controller/selectors';
 import { setActiveRotaryReference } from '../../app/Service';
 import ModeBanner, { BannerType } from '../../modes/ModeBanner';
-import { BPM, PERCENT } from '../../info/units';
+import { BPM, PERCENT } from '../../shared/units';
 import AlarmLimitsModalPopup from './AlarmLimitsModalPopup';
 import ParamAlarms from './ParamAlarms';
