+++ frontend/src/modules/app/screensaver/ScreensaverPage.tsx
@@ -9,7 +9,7 @@ import { useSelector } from 'react-redux';
 import ModeBanner, { BannerType } from '../../modes/ModeBanner';
 import ValueJumbotron from './ValueJumbotron';
 import ControlJumbotron from './ControlJumbotron';
-import { PERCENT, BMIN, LMIN } from '../../info/units';
+import { PERCENT, BMIN, LMIN } from '../../shared/units';
 import {
   getSensorMeasurementsSpO2,
   getCycleMeasurementsRR,
