+++ frontend/src/modules/modals/MultiStepWizard.tsx
@@ -25,7 +25,7 @@ import {
 } from '../../store/controller/selectors';
 import { a11yProps, TabPanel } from '../controllers/TabPanel';
 import ValueInfo from '../dashboard/components/ValueInfo';
-import { BPM, LMIN, PERCENT } from '../info/units';
+import { BPM, LMIN, PERCENT } from '../shared/units';
 import { ParametersRequest, AlarmLimitsRequest, Range } from '../../store/proto/mcu_pb';
 import { MessageType } from '../../store/proto/types';
 import { commitRequest, commitDraftRequest } from '../../store/controller/actions';
